底面、杰尼龟、天空盒

+++ include/Angel.h
@@ -68,7 +68,7 @@ GLuint InitShader( const char* vertexShaderFile,
 const GLfloat  DivideByZeroTolerance = GLfloat(1.0e-07);
 
 //  Degrees-to-radians constant 
-const GLfloat  DegreesToRadians = M_PI / 180.0;
+const GLfloat  DegreesToRadians = (float) (M_PI / 180.0);
 
 }  // namespace Angel
